dashboard(filtering): fold mempool tiles into What's active, graph-only below
